Convert unnecessary floats to int

+++ sixpack/models.py
@@ -205,7 +205,7 @@ class Experiment(object):
 
         redis_results = pipe.execute()
         for idx, k in enumerate(keys):
-            stats[k] = float(redis_results[idx])
+            stats[k] = int(redis_results[idx])
 
         return stats
 
@@ -813,7 +813,7 @@ class Alternative(object):
 
         redis_results = pipe.execute()
         for idx, k in enumerate(keys):
-            stats[k] = float(redis_results[idx])
+            stats[k] = int(redis_results[idx])
 
         return stats
